Refactor builders to new pipeline pattern
