+++ package.json
@@ -29,7 +29,6 @@
     "gulp-inject": "4.2.0",
     "gulp-jscs": "4.0.0",
     "gulp-jshint": "2.0.4",
-    "gulp-load-plugins": "1.5.0",
     "gulp-postcss": "6.3.0",
     "gulp-processhtml": "1.1.0",
     "gulp-rename": "1.2.2",
