+++ tools/misc/arg-filter.js
@@ -1,24 +1,22 @@
 var _ = require('lodash'),
     args = require('yargs').argv,
     combine = require('stream-combiner'),
-    loadGulpPlugins = require('gulp-load-plugins');
+    filter = require('gulp-filter');
 
 module.exports = function argFilter() {
-  var gulpPlugins = loadGulpPlugins(),
-      regexen,
-      pipeline = [];
+  var regexen;
 
   if (!!args.match) {
     regexen = _.map(args.match.split(','), function makeRegex(match) {
       return new RegExp(match.replace('/', '\\/'), 'i');
     });
 
-    pipeline.push(gulpPlugins.filter(function filterFiles(file) {
+    return filter(function filterFiles(file) {
       return _.some(regexen, function testFileWithRegex(regex) {
         return regex.test(file.path);
       });
-    }));
+    });
+  } else {
+    return combine([]);
   }
-
-  return combine(_.compact(pipeline));
 };
